Finished the first menu scene
I've designed and added username capabilities to the menu scene. I've also connected the button so that once the user inputs a valid username, it'll redirect the user to the next scene (which is the character select). The username input has error checks to ensure that the user inputs a valid and non-empty username.